// src/app/core/utils/firebase-code-errors-enum.ts
export enum FirebaseCodeErrorsEnum {
  emailAlreadyExist = 'auth/email-already-exists',
  emailAlreadyInUse = 'auth/email-already-in-use',
  idTokenExpired = 'auth/id-token-expired',
  internalError = 'auth/internal-error',
  invalidDisplayName = 'auth/invalid-display-name',
  invalidEmail = 'auth/invalid-email',
  invalidEmailVerified = 'auth/invalid-email-verified',
  invalidPassword = 'auth/invalid-password',
  wrongPassword = 'auth/wrong-password',
  userNotFound = 'auth/user-not-found',
}